CTime::GetTime assets for valid date time values, so protect against wrong values
It is possible that some of the blogs store data and time which may not match correct value.

I have seen blog info where the date as. Notice that instead of 2020 the entry says 20 which is incorrect. The default values will make sure that information is not lost.

"Wed, 07 Oct 20 00:00:00 +0000"

Note that a better strategy can be devised, but in the meantime protect against wrong values.